Move MockTimeseriesAspectService into its own class
